requirements.txt
beautifulsoup4==4.12.3
deepl==1.20.0
extruct==0.18.0
fastapi==0.115.6
gspread==6.1.4
lxml==5.3.0
zendriver==0.2.3
oauth2client==4.1.3
pydantic==2.10.3
python_dateutil==2.9.0.post0
Requests==2.32.3
tldextract==5.1.3
w3lib==2.2.1
uvicorn==0.34.0